# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SakanMatch API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: listings
    description: Listing operations
  - name: paypal
    description: PayPal subscription operations
  - name: premium
    description: Premium status operations
  - name: storage
    description: Object storage operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get all listings
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingResponse"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - owners only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /listings/my:
    get:
      operationId: getMyListings
      tags: [listings]
      summary: Get current user listings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingResponse"

  /listings/{id}/contact-click:
    post:
      operationId: recordContactClick
      tags: [listings]
      summary: Record a contact click for a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Click recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /paypal/create-subscription:
    post:
      operationId: createSubscription
      tags: [paypal]
      summary: Create a PayPal subscription for premium
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSubscriptionResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Owners only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /paypal/activate-subscription:
    post:
      operationId: activateSubscription
      tags: [paypal]
      summary: Activate a PayPal subscription after approval
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateSubscriptionRequest"
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivateSubscriptionResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /paypal/cancel-subscription:
    post:
      operationId: cancelSubscription
      tags: [paypal]
      summary: Cancel the current user's active PayPal subscription
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: No active subscription to cancel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /paypal/webhook:
    post:
      operationId: paypalWebhook
      tags: [paypal]
      summary: PayPal webhook handler
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /premium/status:
    get:
      operationId: getPremiumStatus
      tags: [premium]
      summary: Get premium status of current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Premium status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremiumStatusResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        role:
          type: string
          enum: [owner, seeker]
      required:
        - email
        - password
        - role

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserResponse"
      required:
        - token
        - user

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [owner, seeker]
        isPremium:
          type: boolean
        premiumActivatedAt:
          type: string
          nullable: true
        premiumSource:
          type: string
          nullable: true
        subscriptionStatus:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - email
        - role
        - isPremium

    CreateListingRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        city:
          type: string
        images:
          type: array
          items:
            type: string
        deposit:
          type: number
          nullable: true
        utilitiesIncluded:
          type: boolean
        furnished:
          type: boolean
        availableFrom:
          type: string
          nullable: true
        roomType:
          type: string
          nullable: true
          enum: [private_room, shared_room, studio, entire_place]
        preferredTenantGender:
          type: string
          enum: [male, female, no_preference]
        minStayMonths:
          type: integer
          nullable: true
        houseRules:
          type: string
          nullable: true
      required:
        - title
        - price
        - city

    ListingResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        city:
          type: string
        images:
          type: array
          items:
            type: string
        deposit:
          type: number
          nullable: true
        utilitiesIncluded:
          type: boolean
        furnished:
          type: boolean
        availableFrom:
          type: string
          nullable: true
        roomType:
          type: string
          nullable: true
          enum: [private_room, shared_room, studio, entire_place]
        preferredTenantGender:
          type: string
          enum: [male, female, no_preference]
        minStayMonths:
          type: integer
          nullable: true
        houseRules:
          type: string
          nullable: true
        isIdentityVerified:
          type: boolean
        isLocationVerified:
          type: boolean
        ownerId:
          type: integer
        ownerEmail:
          type: string
        ownerName:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        viewCount:
          type: integer
          nullable: true
        contactClickCount:
          type: integer
          nullable: true
        requestCount:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - price
        - city
        - ownerId

    CreateSubscriptionResponse:
      type: object
      properties:
        subscriptionID:
          type: string
        approvalUrl:
          type: string
          nullable: true
      required:
        - subscriptionID

    ActivateSubscriptionRequest:
      type: object
      properties:
        subscriptionID:
          type: string
      required:
        - subscriptionID

    ActivateSubscriptionResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
          nullable: true
      required:
        - status

    PremiumStatusResponse:
      type: object
      properties:
        isPremium:
          type: boolean
        premiumActivatedAt:
          type: string
          nullable: true
        premiumSource:
          type: string
          nullable: true
        subscriptionStatus:
          type: string
          nullable: true
        paypalSubscriptionId:
          type: string
          nullable: true
        lastPaymentAt:
          type: string
          nullable: true
      required:
        - isPremium

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath